+++ include/frame.h
@@ -17,8 +17,8 @@
 
 typedef struct {
 	Bitmap bitmap;
-	unsigned long origin_frames;
-	unsigned long usable_frames;
+	size_t origin_frames;
+	size_t usable_frames;
 } FrameAllocator;
 
 extern FrameAllocator frame_allocator;
@@ -27,7 +27,7 @@ extern FrameAllocator frame_allocator;
 void init_frame(void);
 
 /* Allocate memory frame */
-uint64_t alloc_frames(unsigned long count);
+uint64_t alloc_frames(size_t count);
 
 /* Free memory frames */
 void free_frame(uint64_t addr);
